+++ MyCookbook/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-09-20
+### Added
+- Added a dropdown to the recipe category edit field.
+- Added support for middle mouse button click in the recipe browser.
+
+### Changed
+- Changed recipe viewer URL to use recipe ID instead of name.
+
 ## [1.2.1] - 2025-09-06
 ### Fixed
 - Fixed ingredient not being saved when using Escape.